requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
clarifai==2.0.32
click==6.7
configparser==3.5.0
enum34==1.1.6
facebook-sdk==2.0.0
Flask==0.12.2
future==0.15.2
gpsimage==0.0.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.5.1
MarkupSafe==1.0
olefile==0.44
Pillow==4.3.0
PyMySQL==0.7.11
requests==2.18.4
urllib3==1.22
Werkzeug==0.12.2
geopy
